requirements.txt
netmiko>=4.1.2
textfsm>=1.1.3
pyyaml>=6.0.1
rich>=13.7.0 